refactor: fulltext search 기능 제외
