+++ src/main/java/swyp12/team9/server/domain/userlink/repository/UserLinkSearchRepositoryImpl.java
@@ -1,27 +1,23 @@
-package swyp12.team9.server.domain.userlink.repository.search;
-
-import com.querydsl.core.types.dsl.BooleanExpression;
-import com.querydsl.core.types.dsl.StringPath;
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import swyp12.team9.server.domain.userlink.model.UserLink;
-
-import java.util.List;
+package swyp12.team9.server.domain.userlink.repository;
 
 import static swyp12.team9.server.domain.link.model.QLink.link;
 import static swyp12.team9.server.domain.reference.model.QReference.reference;
 import static swyp12.team9.server.domain.referenceuserlink.model.QReferenceUserLink.referenceUserLink;
 import static swyp12.team9.server.domain.userlink.model.QUserLink.userLink;
 
+import com.querydsl.core.types.dsl.BooleanExpression;
+import com.querydsl.core.types.dsl.StringPath;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import swyp12.team9.server.domain.userlink.model.UserLink;
+
 /**
  * UserLink 검색 Repository QueryDSL 구현체
  * <p>
  * 모든 검색 쿼리를 QueryDSL로 통일하여 타입 안전성과 유지보수성을 높입니다.
  * <p>
- * 주요 특징:
- * - 동적 필드 검색: field 파라미터에 따라 검색 대상 필드가 결정됨
- * - 커서 기반 페이징: ID 기반 무한 스크롤 지원
- * - N:N 관계 처리: ReferenceUserLink를 통한 조인
+ * 주요 특징: - 동적 필드 검색: field 파라미터에 따라 검색 대상 필드가 결정됨 - 커서 기반 페이징: ID 기반 무한 스크롤 지원 - N:N 관계 처리: ReferenceUserLink를 통한 조인
  */
 @RequiredArgsConstructor
 public class UserLinkSearchRepositoryImpl implements UserLinkSearchRepositoryCustom {
@@ -62,8 +58,7 @@ public class UserLinkSearchRepositoryImpl implements UserLinkSearchRepositoryCus
     /**
      * 특정 레퍼런스 폴더 내 링크 검색 (커서 기반)
      * <p>
-     * - ReferenceUserLink를 통해 레퍼런스와 연결된 모든 UserLink 검색
-     * - 내 링크뿐 아니라 다른 사용자의 링크도 포함
+     * - ReferenceUserLink를 통해 레퍼런스와 연결된 모든 UserLink 검색 - 내 링크뿐 아니라 다른 사용자의 링크도 포함
      */
     @Override
     public List<UserLink> searchByReference(Long referenceId, String keyword, String field, Long cursorId, int size) {
@@ -86,8 +81,7 @@ public class UserLinkSearchRepositoryImpl implements UserLinkSearchRepositoryCus
     /**
      * 내 모든 레퍼런스 폴더 내 링크 검색 (커서 기반)
      * <p>
-     * - 내가 소유한 모든 레퍼런스 폴더에 속한 UserLink 검색
-     * - DISTINCT로 중복 제거 (하나의 링크가 여러 폴더에 있을 수 있음)
+     * - 내가 소유한 모든 레퍼런스 폴더에 속한 UserLink 검색 - DISTINCT로 중복 제거 (하나의 링크가 여러 폴더에 있을 수 있음)
      */
     @Override
     public List<UserLink> searchAllMyReferences(Long userId, String keyword, String field, Long cursorId, int size) {
@@ -121,8 +115,7 @@ public class UserLinkSearchRepositoryImpl implements UserLinkSearchRepositoryCus
     /**
      * 키워드 검색 조건 (동적 필드 선택)
      * <p>
-     * field가 null이면 모든 필드에서 검색
-     * field가 지정되면 해당 필드에서만 검색
+     * field가 null이면 모든 필드에서 검색 field가 지정되면 해당 필드에서만 검색
      *
      * @param keyword 검색 키워드
      * @param field   검색 필드 (why, memo, title, aiSummary, url)
